리액트 초기설정
