+++ client/src/App.js
@@ -1,0 +1,9 @@
+function App() {
+  return (
+    <>
+      테스트페이지
+    </>
+  );
+}
+
+export default App;
